move shimmers to a diff folder
