--- src/components/Loader.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Loader = () => {
-  return (
-    <div className="loadingio-spinner-dual-ball-di46lniu3nw"><div className="ldio-bqijn9mkrfs">
-<div></div><div></div><div></div>
-</div></div>
-  );
-};
-
-export default Loader;
-
