--- src/components/PokemonsLoader.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const PokemonsLoader = () => {
-  return (
-    <div className='loadingio-spinner-eclipse-pgmernpywjq'>
-      <div className='ldio-iv7gr82qd7r'>
-        <div></div>
-      </div>
-    </div>
-  );
-};
-
-export default PokemonsLoader;
